Move codec into sub-module.
